+++ oddrip/account.go
@@ -17,3 +17,11 @@ func (s *AccountService) GetAPILimits(ctx context.Context) (*types.GetAccountApi
 	}
 	return &out, nil
 }
+
+func (s *AccountService) GetEndpointCosts(ctx context.Context) (*types.GetAccountEndpointCostsResponse, error) {
+	var out types.GetAccountEndpointCostsResponse
+	if err := s.client.get(ctx, joinPath("account", "endpoint_costs"), nil, &out); err != nil {
+		return nil, err
+	}
+	return &out, nil
+}